third_party/WebKit/LayoutTests/virtual/implsidepainting/inspector/timeline/timeline-frames-expected.txt

Tests reporting and canceling of didBeginFrame events in Timeline

Frame 0
TimeStamp : FRAME 1
Frame 1
TimeStamp : FRAME 2
TimeStamp : FRAME 3
Frame 2
TimeStamp : FRAME 4
